requirements.txt
aiogram==3.1.0b7
aiofiles
python-dotenv
python-multipart
